ReviewsReply_Agent/requirements.txt
python-docx
python-dotenv
Flask
requests
